Move `StructuralQuery` to its own module
