+++ packages/graph/hash_graph/bench/benches/representative_read/ontology/entity_type.rs
@@ -1,8 +1,7 @@
 use criterion::{BatchSize::SmallInput, Bencher};
 use graph::{
-    shared::subgraph::depths::GraphResolveDepths,
+    shared::subgraph::{depths::GraphResolveDepths, query::StructuralQuery},
     store::{query::Filter, EntityTypeStore},
-    subgraph::StructuralQuery,
 };
 use rand::{prelude::IteratorRandom, thread_rng};
 use tokio::runtime::Runtime;
