+++ packages/graph/hash_graph/lib/graph/src/shared/subgraph/mod.rs
@@ -1,23 +1,19 @@
 use std::{
     collections::{HashMap, HashSet},
-    fmt::{Debug, Formatter},
+    fmt::Debug,
 };
 
 use depths::GraphResolveDepths;
 use edges::Edges;
-use serde::{Deserialize, Serialize};
-use type_system::{DataType, EntityType, PropertyType};
+use serde::Serialize;
 use utoipa::ToSchema;
 use vertices::Vertex;
 
-use crate::{
-    knowledge::EntityProperties,
-    shared::identifier::GraphElementId,
-    store::query::{Filter, QueryRecord},
-};
+use crate::shared::identifier::GraphElementId;
 
 mod depths;
 mod edges;
+mod query;
 mod vertices;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
@@ -49,35 +45,5 @@ impl Extend<Self> for Subgraph {
             self.vertices.extend(subgraph.vertices.into_iter());
             self.edges.extend(subgraph.edges.into_iter());
         }
-    }
-}
-
-/// A [`Filter`] to query the datastore, recursively resolving according to the
-/// [`GraphResolveDepths`].
-#[derive(Deserialize, ToSchema)]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-#[aliases(
-    DataTypeStructuralQuery = StructuralQuery<'static, DataType>,
-    PropertyTypeStructuralQuery = StructuralQuery<'static, PropertyType>,
-    EntityTypeStructuralQuery = StructuralQuery<'static, EntityType>,
-    EntityStructuralQuery = StructuralQuery<'static, EntityProperties>,
-)]
-pub struct StructuralQuery<'q, T: QueryRecord> {
-    #[serde(bound = "'de: 'q, T::Path<'q>: Deserialize<'de>")]
-    pub filter: Filter<'q, T>,
-    pub graph_resolve_depths: GraphResolveDepths,
-}
-
-// TODO: Derive traits when bounds are generated correctly
-//   see https://github.com/rust-lang/rust/issues/26925
-impl<'q, T> Debug for StructuralQuery<'q, T>
-where
-    T: QueryRecord<Path<'q>: Debug>,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("StructuralQuery")
-            .field("filter", &self.filter)
-            .field("graph_resolve_depths", &self.graph_resolve_depths)
-            .finish()
     }
 }
