+++ packages/graph/hash_graph/lib/graph/src/shared/subgraph/query.rs
@@ -1,0 +1,41 @@
+use std::fmt::{Debug, Formatter};
+
+use serde::Deserialize;
+use type_system::{DataType, EntityType, PropertyType};
+use utoipa::ToSchema;
+
+use crate::{
+    knowledge::EntityProperties,
+    store::query::{Filter, QueryRecord},
+    subgraph::depths::GraphResolveDepths,
+};
+
+/// A [`Filter`] to query the datastore, recursively resolving according to the
+/// [`GraphResolveDepths`].
+#[derive(Deserialize, ToSchema)]
+#[serde(deny_unknown_fields, rename_all = "camelCase")]
+#[aliases(
+    DataTypeStructuralQuery = StructuralQuery<'static, DataType>,
+    PropertyTypeStructuralQuery = StructuralQuery<'static, PropertyType>,
+    EntityTypeStructuralQuery = StructuralQuery<'static, EntityType>,
+    EntityStructuralQuery = StructuralQuery<'static, EntityProperties>,
+)]
+pub struct StructuralQuery<'q, T: QueryRecord> {
+    #[serde(bound = "'de: 'q, T::Path<'q>: Deserialize<'de>")]
+    pub filter: Filter<'q, T>,
+    pub graph_resolve_depths: GraphResolveDepths,
+}
+
+// TODO: Derive traits when bounds are generated correctly
+//   see https://github.com/rust-lang/rust/issues/26925
+impl<'q, T> Debug for StructuralQuery<'q, T>
+where
+    T: QueryRecord<Path<'q>: Debug>,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("StructuralQuery")
+            .field("filter", &self.filter)
+            .field("graph_resolve_depths", &self.graph_resolve_depths)
+            .finish()
+    }
+}
